Add shared ExpressionEditor component for advanced filter mode in FilterSettings
